+++ test/test_links.py
@@ -73,3 +73,23 @@ def test_link_alt_text(tmp_path):
         alt_text="GitHub repository of the fpdf2 library",
     )
     assert_pdf_equal(pdf, HERE / "link_alt_text.pdf", tmp_path)
+
+
+def test_link_with_zoom(tmp_path):
+    pdf = FPDF()
+    pdf.set_font("helvetica", size=24)
+    pdf.add_page()
+    link = pdf.add_link()
+    pdf.set_link(link, page=2, zoom=4)
+    pdf.set_xy(50, 50)
+    pdf.cell(w=100, h=10, txt="Link to 2nd page zoomed", border=1, align="C", link=link)
+    pdf.add_page()
+    pdf.multi_cell(
+        pdf.epw,
+        txt="Lorem ipsum dolor sit amet, consectetur adipiscing elit,"
+        " sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+        " Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."
+        " Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur."
+        " Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+    )
+    assert_pdf_equal(pdf, HERE / "link_with_zoom.pdf", tmp_path)